requirements.txt
requests
beautifulsoup4
pandas
google-cloud-storage
Flask
